+++ internal/provider/data_source_nxos_access_list_test.go
@@ -72,7 +72,9 @@ func TestAccDataSourceNxosAccessList(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("data.nxos_access_list.test", "access_lists.ACL1.entries.10.telemetry_queue", "false"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.nxos_access_list.test", "access_lists.ACL1.entries.10.type_of_service", "0"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.nxos_access_list.test", "ingress_interfaces.eth1/10.access_list_name", "ACL1"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.nxos_access_list.test", "ingress_vty_access_list_name", "ACL1"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.nxos_access_list.test", "egress_interfaces.eth1/10.access_list_name", "ACL1"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.nxos_access_list.test", "egress_vty_access_list_name", "ACL1"))
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
@@ -157,11 +159,13 @@ func testAccDataSourceNxosAccessListConfig() string {
 	config += `			access_list_name = "ACL1"` + "\n"
 	config += `		}` + "\n"
 	config += `	}` + "\n"
+	config += `	ingress_vty_access_list_name = "ACL1"` + "\n"
 	config += `	egress_interfaces = {` + "\n"
 	config += `		"eth1/10" = {` + "\n"
 	config += `			access_list_name = "ACL1"` + "\n"
 	config += `		}` + "\n"
 	config += `	}` + "\n"
+	config += `	egress_vty_access_list_name = "ACL1"` + "\n"
 	config += `	depends_on = [nxos_dme.PreReq0, ]` + "\n"
 	config += `}` + "\n"