+++ internal/provider/resource_nxos_system_test.go
@@ -166,6 +166,9 @@ func TestAccNxosSystem(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("nxos_system.test", "lldp_interfaces.eth1/1.tlv_vlan", "100"))
 	checks = append(checks, resource.TestCheckResourceAttr("nxos_system.test", "copp_admin_state", "enabled"))
 	checks = append(checks, resource.TestCheckResourceAttr("nxos_system.test", "copp_rate_limiter", "true"))
+	checks = append(checks, resource.TestCheckResourceAttr("nxos_system.test", "console_exec_timeout", "30"))
+	checks = append(checks, resource.TestCheckResourceAttr("nxos_system.test", "vty_exec_timeout", "30"))
+	checks = append(checks, resource.TestCheckResourceAttr("nxos_system.test", "vty_session_limit", "16"))
 	var tfVersion *goversion.Version
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
@@ -396,6 +399,9 @@ func testAccNxosSystemConfig_all() string {
 	config += `	}` + "\n"
 	config += `	copp_admin_state = "enabled"` + "\n"
 	config += `	copp_rate_limiter = true` + "\n"
+	config += `	console_exec_timeout = 30` + "\n"
+	config += `	vty_exec_timeout = 30` + "\n"
+	config += `	vty_session_limit = 16` + "\n"
 	config += `	depends_on = [nxos_dme.PreReq0, nxos_dme.PreReq1, ]` + "\n"
 	config += `}` + "\n"
 	return config